requirements.txt
torch
torchvision
diffusers
transformers
pytorch-lightning
accelerate
tqdm
Pillow

